+++ src/layer/cesium/vectorContext.ts
@@ -30,7 +30,7 @@ import {
   primitives as primitivesSymbol,
   scaleSymbol,
 } from '../vectorSymbols.js';
-import PanoramaMap from '../../map/panoramaMap.js';
+import type PanoramaMap from '../../map/panoramaMap.js';
 import { getResolution } from '../../map/cesiumMapHelpers.js';
 
 export function setReferenceForPicking(
@@ -412,6 +412,9 @@ export default class VectorContext implements CesiumVectorContext {
     this.scaledPrimitives.removeAll();
     this.billboards.removeAll();
     this.labels.removeAll();
+    for (const feature of this._featureItems.keys()) {
+      feature[primitivesSymbol] = undefined;
+    }
     this._featureItems.clear();
     this._convertingFeatures.forEach((destroy) => {
       destroy(true);
@@ -436,6 +439,9 @@ export default class VectorContext implements CesiumVectorContext {
       destroy(true);
     });
     this._convertingFeatures.clear();
+    for (const feature of this._featureItems.keys()) {
+      feature[primitivesSymbol] = undefined;
+    }
     this._featureItems.clear();
     this._postRenderListener();
   }